move TestClock to mock package as MockClock
