+++ src/test/java/com/team766/framework/TimedPredicateTest.java
@@ -2,7 +2,7 @@ package com.team766.framework;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.team766.hal.TestClock;
+import com.team766.hal.mock.MockClock;
 import java.util.function.BooleanSupplier;
 import org.junit.jupiter.api.Test;
 
@@ -10,7 +10,7 @@ public class TimedPredicateTest {
 
     @Test
     public void testTimedPredicateTimedOut() {
-        TestClock testClock = new TestClock(1710411240.0);
+        MockClock testClock = new MockClock(1710411240.0);
         Context.TimedPredicate predicate =
                 new Context.TimedPredicate(testClock, () -> false, 1.766);
         assertFalse(predicate.getAsBoolean());
@@ -23,7 +23,7 @@ public class TimedPredicateTest {
 
     @Test
     public void testTimedPredicateCondition() {
-        TestClock testClock = new TestClock(1710411240.0);
+        MockClock testClock = new MockClock(1710411240.0);
         Context.TimedPredicate predicate =
                 new Context.TimedPredicate(
                         testClock,
